Revert "MemoryIntArray: don't double close."
The reverted change causes a regression where we can get an IllegalStateException
during finalization as we are adopting the native fd in a ParcelFileDescriptor
which takes ownership of the fd. However, the order of finalization is undefined
and if the ParcelFileDescriptor is finalized before the MemoryIntArray we would
get an exception when running the finalization of the latter.

This reverts commit 993054d90c043de494b43b913a690f31b600b4f1.

bug:124056170