Take snapshot directly during change transition
Using the tasksnapshotcontroller here is actually wrong... the
snapshot being created is already out-of-date and could have
scaling/other things applied to it. Additionally, it can clash
with things listening for task-snapshot changes which is also
wrong.

This changes to getting a snapshot directly which is more
correct and faster (since it doesn't go through the snapshot
controller's caching thing).

Bug: 132783168
Test: Manual, change transitions still work. android on chrome
      doesn't crash on tablet-mode switch

Change-Id: I726ed174765ac64bedb7cd130719190a06ad42ab