Fix notification shade inefficient contrast
Because the color of the notification shade contains alpha channel,
the colors in wallpaper blend with the color of the notification
shade. After the blending with wallpaper colors and notification
shade color, the contrast ratio between gear icon and the blending
color is not enough.
The solution is to use the solid color rather than the colors with
the transparency.
Light Theme: shade=#E8EAED gear=#5f6368
Dark Theme: shade=#202124 gear=#9aa0a6

Fixes: 124379911
Test: atest SystemUITests
Change-Id: I0e436a1f6b5507c5083b4b1b9811166d0fb41da7