Fix a bug about missing committing app visibility
If we have pending transitions, committing visibility would be delayed
until the opening apps are all drawn. However, the original logic in
AppWindowToken.transferStartingWindow would just remove the app from
mOpeningApps which caused the app visibility would not be committed.

This CL doesn't remove the app from mOpeningApps while transferring the
starting window. Instead, we set a flag to indicate that we no longer
need to apply an animation to the app. In this way, the app visibility
will be committed.

Fix: 131180307
Test: atest WindowManagerSmokeTest
Test: Steps in the bug
Change-Id: Ia84d73136dcbda07ae055b91f5a038e1bc070d2a