Not rendering the dark icons with the shelf anymore
Instead it is rendered by a dedicated container.
This will help with a lot of the transitions that
are necessary for the bypass and also fixes the
coloring of the notification icons on the lockscreen.

Fixes: 133869048
Bug: 130327302
Bug: 133182504
Test: atest SystemUITests
Change-Id: I422b3da488fdac6577173d41b26c67f81c7826c5
(cherry picked from commit c3e3d6d762b9c9de800f2c85b268d49850f7e490)