Do not notify SS change with invalid subId
If a service state change is reported as a subscription is becoming
invalid (i.e. no longer connected to a phone), the current behavior can
sometimes notify this change to all PhoneStateListeners regardless of
which subId they registered to listen on (namely this happens when the
subscription was attached to the default phone).

This change removes that behavior so that no listeners are notified if
the subscription is invalid.

Test: manually verified on DSDS that when
  1. erasing eSIM when it is default and
  2. removing pSIM when it is default,
  the service state triangle for the other SIM is unaffected
Bug: 133140128
Change-Id: I906147b2d13be0bc26f8d31ec361fcd016c2cda0