Only consider gone for layout if parent is gone for layout
If we check getParentWindowHidden, that determines mostly actual
visibility. However, we don't want that because we still would
like to follow the parent's window layout lifecycle, as otherwise
we may be stuck in a transition in case the parent window is
hidden but the child is waiting for a layout to happen.

Test: Click "Customize" on wallpaper picker, go back, ensure no
transition timeout
Fixes: 135976008

Change-Id: I66aeab29a81cd82b170aaf337249616b1f559848
(cherry picked from commit 07b079d774ce5d181aa1f1f448ad9e11df603cb1)